Extract backgroundable classes in the scope of frontend actions
