+++ frontend/actions/src/main/java/com/virtuslab/gitmachete/frontend/actions/backgroundables/FetchBackgroundable.java
@@ -11,7 +11,6 @@ import git4idea.repo.GitRepository;
 import lombok.CustomLog;
 import lombok.val;
 import org.checkerframework.checker.guieffect.qual.UIEffect;
-import org.checkerframework.checker.nullness.qual.Nullable;
 import org.checkerframework.checker.tainting.qual.Untainted;
 
 import com.virtuslab.qual.guieffect.UIThreadUnsafe;
@@ -25,7 +24,6 @@ public class FetchBackgroundable extends Task.Backgroundable {
   private final String refspec;
   private final @Untainted String failureNotificationText;
   private final String successNotificationText;
-  private final @Nullable @Untainted String taskSubtitle;
 
   /** Use as {@code remoteName} when referring to the local repository. */
   public static final String LOCAL_REPOSITORY_NAME = ".";
@@ -35,7 +33,6 @@ public class FetchBackgroundable extends Task.Backgroundable {
       String remoteName,
       String refspec,
       String taskTitle,
-      @Nullable @Untainted String taskSubtitle,
       @Untainted String failureNotificationText,
       String successNotificationText) {
     super(project, taskTitle, /* canBeCancelled */ true);
@@ -43,29 +40,13 @@ public class FetchBackgroundable extends Task.Backgroundable {
     this.gitRepository = gitRepository;
     this.remoteName = remoteName;
     this.refspec = refspec;
-    this.taskSubtitle = taskSubtitle;
     this.failureNotificationText = failureNotificationText;
     this.successNotificationText = successNotificationText;
-  }
-
-  public FetchBackgroundable(Project project,
-      GitRepository gitRepository,
-      String remoteName,
-      String refspec,
-      String taskTitle,
-      @Untainted String failureNotificationText,
-      String successNotificationText) {
-    this(project, gitRepository, remoteName, refspec, taskTitle, /* taskSubtitle */ null, failureNotificationText,
-        successNotificationText);
   }
 
   @Override
   @UIThreadUnsafe
   public void run(ProgressIndicator indicator) {
-    if (taskSubtitle != null) {
-      // This method set a text under a progress bar (despite docstring)
-      indicator.setText(taskSubtitle);
-    }
     val fetchSupport = GitFetchSupport.fetchSupport(project);
     GitRemote remote = remoteName.equals(LOCAL_REPOSITORY_NAME)
         ? GitRemote.DOT
