+++ frontend/actions/src/main/java/com/virtuslab/gitmachete/frontend/actions/base/BaseSyncToParentByMergeAction.java
@@ -23,8 +23,8 @@ import org.checkerframework.checker.i18nformatter.qual.I18nFormat;
 import org.checkerframework.checker.tainting.qual.Untainted;
 
 import com.virtuslab.gitmachete.backend.api.SyncToParentStatus;
+import com.virtuslab.gitmachete.frontend.actions.backgroundables.CheckoutBackgroundable;
 import com.virtuslab.gitmachete.frontend.actions.common.MergeProps;
-import com.virtuslab.gitmachete.frontend.actions.contextmenu.CheckoutSelectedAction;
 import com.virtuslab.gitmachete.frontend.defs.ActionPlaces;
 import com.virtuslab.qual.guieffect.UIThreadUnsafe;
 
@@ -136,7 +136,7 @@ public abstract class BaseSyncToParentByMergeAction extends BaseGitMacheteReposi
       public void run(ProgressIndicator indicator) {
         LOG.info("Checking out '${movingBranch}' branch and merging ${stayingBranch} into it");
 
-        CheckoutSelectedAction.doCheckout(project, indicator, movingBranch, gitRepository);
+        CheckoutBackgroundable.doCheckout(project, indicator, movingBranch, gitRepository);
         GitBrancher.getInstance(project)
             .merge(stayingBranch, GitBrancher.DeleteOnMergeOption.NOTHING, Collections.singletonList(gitRepository));
       }
