+++ frontend/actions/src/main/java/com/virtuslab/gitmachete/frontend/actions/common/FastForwardMerge.java
@@ -48,12 +48,11 @@ public final class FastForwardMerge {
         LOCAL_REPOSITORY_NAME,
         refspecFromChildToParent,
         getString("action.GitMachete.BaseFastForwardMergeToParentAction.task-title"),
-        getNonHtmlString("action.GitMachete.BaseFastForwardMergeToParentAction.task-subtitle"),
         fetchNotificationTextPrefix + failFFMergeNotification,
         fetchNotificationTextPrefix + successFFMergeNotification).queue();
   }
 
-  public static void perform(Project project,
+  public static CheckRemoteBranchBackgroundable createBackgroundable(Project project,
       GitRepository gitRepository,
       MergeProps mergeProps,
       @Untainted String fetchNotificationTextPrefix) {
@@ -61,7 +60,7 @@ public final class FastForwardMerge {
     val currentBranchName = Option.of(gitRepository.getCurrentBranch()).map(GitReference::getName).getOrNull();
     val failFfMergeNotificationTitle = getString(
         "action.GitMachete.BaseFastForwardMergeToParentAction.notification.title.ff-fail");
-    new CheckRemoteBranchBackgroundable(project, gitRepository, stayingName, failFfMergeNotificationTitle,
+    return new CheckRemoteBranchBackgroundable(project, gitRepository, stayingName, failFfMergeNotificationTitle,
         fetchNotificationTextPrefix) {
       @Override
       @UIEffect
@@ -72,6 +71,6 @@ public final class FastForwardMerge {
           mergeNonCurrentBranch(project, gitRepository, mergeProps, fetchNotificationTextPrefix);
         }
       }
-    }.queue();
+    };
   }
 }
