+++ frontend/resourcebundles/src/main/resources/GitMacheteBundle.properties
@@ -11,7 +11,7 @@
 # - "Cannot create file 'machete': Directory with the same name exists"
 # because it is never possible while the directory exists
 # - "Could not open Git Machete tab"
-# because we are not able to specify why it was not opened; it could be missing tool window or missing content(tab)
+# because we are not able to specify why it was not opened; it could be a missing tool window or missing content(tab)
 # but we neither want to expose the technical details nor we know the actual reason
 # why the tool window or the content tab is null; most likely it's disabled Git VCS
 
@@ -52,8 +52,7 @@ action.GitMachete.BaseFastForwardMergeToParentAction.description=Fast-forward me
 action.GitMachete.BaseFastForwardMergeToParentAction.notification.title.ff-fail=Fast-forward merge failed
 action.GitMachete.BaseFastForwardMergeToParentAction.notification.text.ff-fail=Fast-forward merge branch ''{1}'' to ''{0}'' failed.
 action.GitMachete.BaseFastForwardMergeToParentAction.notification.text.ff-success=Branch ''{1}'' has been fast-forward merged to ''{0}''.
-action.GitMachete.BaseFastForwardMergeToParentAction.task-title=Fast-forward merging
-action.GitMachete.BaseFastForwardMergeToParentAction.task-subtitle=Fast-forwarding\u2026
+action.GitMachete.BaseFastForwardMergeToParentAction.task-title=Fast-forwarding\u2026
 
 
 action.GitMachete.BaseSyncToParentByMergeAction.description-action-name=Merge parent into branch
@@ -75,6 +74,7 @@ action.GitMachete.BaseOverrideForkPointAction.dialog.override-fork-point.title=O
 action.GitMachete.BaseOverrideForkPointAction.dialog.override-fork-point.radio-button.parent=Parent branch ({0})
 action.GitMachete.BaseOverrideForkPointAction.dialog.override-fork-point.radio-button.inferred=Inferred commit ({0})
 action.GitMachete.BaseOverrideForkPointAction.dialog.override-fork-point.radio-button.custom=Manually picked commit:
+action.GitMachete.BaseOverrideForkPointAction.task.title=Overriding fork point\u2026
 
 
 action.GitMachete.BasePullAction.action-name=P_ull
@@ -103,7 +103,6 @@ action.GitMachete.BaseSyncToParentByRebaseAction.description.repository.state.on
 action.GitMachete.BaseSyncToParentByRebaseAction.hook.task-title=Running machete-pre-rebase hook
 action.GitMachete.BaseSyncToParentByRebaseAction.notification.title.rebase-abort=Rebase aborted
 action.GitMachete.BaseSyncToParentByRebaseAction.notification.title.rebase-fail=Rebase failed
-action.GitMachete.BaseSyncToParentByRebaseAction.task-title=Rebasing
 action.GitMachete.BaseSyncToParentByRebaseAction.text=Sync to Parent by _Rebase\u2026
 
 
@@ -120,11 +119,10 @@ action.GitMachete.BaseResetToRemoteAction.info-dialog.message.HTML=\
     <tt>git reset --keep {2}</tt><br><br> \
     Do you want to proceed?\
   </html>
-action.GitMachete.BaseResetToRemoteAction.notification.title=Resetting
+action.GitMachete.BaseResetToRemoteAction.notification.title=Git Machete Reset
 action.GitMachete.BaseResetToRemoteAction.notification.title.reset-fail=Reset of branch ''{0}'' failed
 action.GitMachete.BaseResetToRemoteAction.notification.title.reset-success.HTML=<html>Branch <b>{0}</b> has been reset to remote</html>
-action.GitMachete.BaseResetToRemoteAction.task-title=Resetting branch
-action.GitMachete.BaseResetToRemoteAction.task-subtitle=Resetting\u2026
+action.GitMachete.BaseResetToRemoteAction.task-title=Resetting\u2026
 
 
 action.GitMachete.BaseSlideInBelowAction.description=Slide in branch below ''{0}''
@@ -147,6 +145,7 @@ action.GitMachete.BaseSlideInBelowAction.notification.message.multiple-remotes=M
 action.GitMachete.BaseSlideInBelowAction.notification.message.slide-in-under-itself-or-its-descendant=Cannot slide in the branch under itself or any of its descendants
 
 
+
 action.GitMachete.BaseSlideOutAction.deletion-suggestion-dialog.title=Git Machete Branch Deletion Suggestion
 action.GitMachete.BaseSlideOutAction.deletion-suggestion-dialog.delete-text=Slide Out \\& &Delete Local Branch
 action.GitMachete.BaseSlideOutAction.deletion-suggestion-dialog.cancel-text=&Cancel (do not slide out)
@@ -162,6 +161,7 @@ action.GitMachete.BaseSlideOutAction.notification.title.slide-out-info.canceled=
 action.GitMachete.BaseSlideOutAction.notification.title.slide-out-success.of-current.HTML=<html>Branch <b>{0}</b> has been slid out but not deleted as it is currently checked out</html>
 action.GitMachete.BaseSlideOutAction.notification.title.slide-out-success.with-delete.HTML=<html>Branch <b>{0}</b> has been slid out and deleted</html>
 action.GitMachete.BaseSlideOutAction.notification.title.slide-out-success.without-delete.HTML=<html>Branch <b>{0}</b> has been slid out but not deleted</html>
+action.GitMachete.BaseSlideOutAction.task.title=Deleting branch if required\u2026
 
 
 action.GitMachete.BaseSquashAction.dialog.title=Squashed Commit Message
@@ -201,7 +201,7 @@ action.GitMachete.CheckoutSelectedAction.undefined.branch-name=Checkout disabled
 
 action.GitMachete.BaseCheckoutAction.description.disabled.currently-checked-out=Branch ''{0}'' is currently checked out
 action.GitMachete.BaseCheckoutAction.description.precise=Checkout branch ''{0}''
-action.GitMachete.BaseCheckoutAction.task-title=Checking out
+action.GitMachete.BaseCheckoutAction.task-title=Checking out\u2026
 action.GitMachete.BaseCheckoutAction.undefined.current-branch=There is no current branch; the repository is probably in detached HEAD state
 
 
@@ -298,7 +298,7 @@ string.GitMachete.BranchOrCommitCellRendererComponent.sync-to-parent-status-tool
 
 string.GitMachete.BranchOrCommitCellRendererComponent.sync-to-remote-status-text.ahead-of-remote=(ahead of {0})
 string.GitMachete.BranchOrCommitCellRendererComponent.sync-to-remote-status-text.behind-remote=(behind {0})
-# To avoid clutter we omit `& newer than` part in status label, coz this is default situation
+# To avoid clutter, we omit `& newer than` part in status label, coz this is default situation
 string.GitMachete.BranchOrCommitCellRendererComponent.sync-to-remote-status-text.diverged-from-and-newer-than-remote=(diverged from {0})
 string.GitMachete.BranchOrCommitCellRendererComponent.sync-to-remote-status-text.diverged-from-and-older-than-remote=(diverged from & older than {0})
 string.GitMachete.BranchOrCommitCellRendererComponent.sync-to-remote-status-text.untracked=(untracked)
